requirements.txt
numpy
negspy
dask
requests
h5py
pandas
slugid
sortedcontainers
nose
pybbi>=0.2.0
Click>=7
pydantic
typing
biopython
